backend/requirements.txt
alembic==1.17.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.29.0
billiard==4.2.2
black==23.12.1
boto3==1.40.70
botocore==1.40.70
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
fastapi==0.104.1
filelock==3.20.0
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.3.0
isort==5.13.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
mypy==1.18.2
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pathspec==0.12.1
pgvector==0.2.5
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pytest==7.4.4
python-dateutil==2.9.0.post0
python-multipart==0.0.6
pdfminer.six==20240706
python-docx==1.1.2
PyYAML==6.0.3
redis==5.3.1
regex==2025.11.3
requests==2.32.5
ruff==0.1.15
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==2.7.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.27.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.7.1
tqdm==4.67.1
transformers==4.57.1
triton==3.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0.post1
vine==5.1.0
wcwidth==0.2.14
openai==1.59.3
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
reportlab==4.0.9
